User 의존성 주입
1. 기존 구조를 버리고 UserController와 UserService를 만들어 의존성 주입을 추가
