+++ backend/src/controller/user.controller.ts
@@ -1,58 +1,12 @@
-import dayjs from "dayjs";
-import randomString from "randomstring";
-import { Op, Transaction } from "sequelize";
+import { Transaction } from "sequelize";
 import { File } from "formidable";
 
 import logger from "../logger/logger";
-import { deleteFile, isDefaultFile, uploadFile } from "../util/firebase";
-import { createDigest } from "../util/password";
 
 import sequelize from "../model";
 import { User, ICreate, IUpdate, IUserResponse } from "../model/user.model";
-import { UserRole } from "../model/userRole.model";
 
-import NotFoundError from "../error/notFound";
-import ForbiddenError from "../error/forbidden";
-import UnauthorizedError from "../error/unauthorized";
-import ConflictError from "../error/conflict";
 import UserService from "../service/user.service";
-
-const FOLDER_NAME = "users";
-
-const createCode = async (): Promise<string> => {
-    let isNot = true;
-    let code = "";
-
-    while (isNot) {
-        code = randomString.generate({
-            length: 6,
-            charset: "alphanumeric"
-        });
-
-        const user: User | null = await User.findOne({
-            where: { code }
-        });
-
-        if (!user) isNot = false;
-    }
-
-    return code;
-};
-
-const createProfilePath = (userId: number, file: File): string | null => {
-    let path: string | null = "";
-    const reqFileName = file.originalFilename!;
-    const isDefault = isDefaultFile(reqFileName);
-
-    /**
-     * Frontend에선 static으로 default.jpg,png,svg 셋 중 하나 갖고있다가
-     * 사용자가 profile을 내리면 그걸로 넣고 요청
-     */
-    if (isDefault) path = null;
-    else path = `${FOLDER_NAME}/${userId}/profile/${dayjs().valueOf()}.${reqFileName}`;
-
-    return path;
-};
 
 class UserController {
     private userService: UserService;
@@ -66,138 +20,50 @@ class UserController {
 
         return result;
     };
-}
 
-export const controller = {
-    /**
-     * 유저와 커플의 정보를 가져옵니다.
-     * @param userId User Id
-     * @returns A {@link IUserResponse}
-     */
-    getUsers: async (userId: number): Promise<IUserResponse> => {
-        const userService = new UserService();
-        const result: IUserResponse = await userService.getUser(userId);
-
-        return result;
-    },
-    /**
-     * 유저 정보를 생성합니다.
-     * @param data A {@link ICreate}
-     */
-    createUser: async (data: ICreate): Promise<void> => {
-        const user: User | null = await User.findOne({
-            where: {
-                [Op.or]: [{ email: data.email }, { phone: data.phone }]
-            }
-        });
-
-        if (user) throw new ConflictError("Duplicated User");
-
-        const transaction = await sequelize.transaction();
-        const hash: string = await createDigest(data.password);
-        data.password = hash;
-        data.code = await createCode();
+    createUser = async (data: ICreate): Promise<void> => {
+        const transaction: Transaction = await sequelize.transaction();
 
         try {
-            const createdUser: User = await User.create(
-                {
-                    snsId: data.snsId,
-                    code: data.code,
-                    name: data.name,
-                    email: data.email,
-                    birthday: new Date(data.birthday),
-                    password: hash,
-                    phone: data.phone,
-                    eventNofi: data.eventNofi
-                },
-                { transaction }
-            );
-
-            await UserRole.create(
-                {
-                    userId: createdUser.userId,
-                    roleId: 4
-                },
-                { transaction }
-            );
-
-            transaction.commit();
-        } catch (error) {
-            transaction.rollback();
-            throw error;
-        }
-
-        logger.debug(`Created User => ${data.email}`);
-    },
-    /**
-     * 유저의 정보를 수정합니다.
-     * @param data A {@link IUpdate}
-     * @param profile User Profile
-     */
-    updateUser: async (data: IUpdate, file?: File): Promise<void> => {
-        let isUpload = false;
-        let path: string | null = "";
-
-        const user: User | null = await User.findOne({
-            where: { userId: data.userId }
-        });
-
-        if (!user) throw new NotFoundError("Not Found User");
-        else if (user.deleted) throw new ForbiddenError("User is deleted");
-
-        let prevProfile: string | null = user.profile;
-        let transaction: Transaction | undefined = undefined;
-
-        try {
-            transaction = await sequelize.transaction();
-
-            if (file) {
-                data.profile = createProfilePath(data.userId, file);
-
-                // profile 있으면 업로드
-                if (data.profile) {
-                    await uploadFile(data.profile, file.filepath);
-                    isUpload = true;
-
-                    if (prevProfile) await deleteFile(prevProfile); // 전에 있던 profile 삭제
-                } else if (prevProfile && !data.profile) {
-                    // default 이미지로 변경시
-                    await deleteFile(prevProfile);
-                }
-            }
-
-            await user.update(data);
-            logger.debug(`Update Data => ${JSON.stringify(data)}`);
-
+            await this.userService.createUser(transaction, data);
             await transaction.commit();
         } catch (error) {
-            // Firebase에는 업로드 되었지만 DB 오류가 발생했다면 Firebase Profile 삭제
-            if (data.profile && isUpload) {
-                await deleteFile(path!);
-                logger.error(`After updating the firebase, a db error occurred and the firebase profile is deleted => ${path}`);
-            }
+            await transaction.rollback();
+            logger.error(`User create error => ${JSON.stringify(error)}`);
 
-            if (transaction) await transaction.rollback();
             throw error;
         }
-    },
-    /**
-     * 사용자 정보 삭제이며, Couple이 있는 경우 Frontend에서 연인 끊기 후 삭제를 요청.
-     * @param userId User Id
-     */
-    deleteUser: async (userId: number): Promise<void> => {
-        const user: User | null = await User.findOne({ where: { userId: userId } });
+    };
 
-        if (!user) throw new NotFoundError("Not Found User");
+    updateUser = async (data: IUpdate, file?: File): Promise<User> => {
+        const transaction = await sequelize.transaction();
 
-        await user.update({
-            deleted: true,
-            deletedTime: new Date(dayjs().valueOf())
-        });
+        try {
+            const updateUser: User = await this.userService.updateUser(transaction, data, file);
+            await transaction.commit();
 
-        if (user.profile) await deleteFile(user.profile);
-        logger.debug(`Success Deleted userId => ${userId}`);
-    }
-};
+            return updateUser;
+        } catch (error) {
+            await transaction.rollback();
+            logger.error(`User update error => ${JSON.stringify(error)}`);
 
-export const controller2 = UserController;
+            throw error;
+        }
+    };
+
+    deleteUser = async (userId: number): Promise<void> => {
+        const transaction = await sequelize.transaction();
+
+        try {
+            await this.userService.deleteUser(transaction, userId);
+            await transaction.commit();
+        } catch (error) {
+            await transaction.rollback();
+            logger.error(`User delete error => ${JSON.stringify(error)}`);
+
+            throw error;
+        }
+    };
+}
+
+export default UserController;
