+++ backend/src/service/user.service.ts
@@ -1,16 +1,64 @@
+import dayjs from "dayjs";
+import { File } from "formidable";
+import randomString from "randomstring";
 import { Op, Transaction } from "sequelize";
 
-import sequelize from "../model";
-import { User, IUserResponse } from "../model/user.model";
+import logger from "../logger/logger";
+import { UserRole } from "../model/userRole.model";
+import { User, IUserResponse, ICreate, IUpdate } from "../model/user.model";
 
 import UnauthorizedError from "../error/unauthorized";
+import ConflictError from "../error/conflict";
+
+import { deleteFile, isDefaultFile, uploadFile } from "../util/firebase";
+import { createDigest } from "../util/password";
+import NotFoundError from "../error/notFound";
+import ForbiddenError from "../error/forbidden";
 
 class UserService {
+    private FOLDER_NAME = "users";
+
+    private createCode = async (): Promise<string> => {
+        let isNot = true;
+        let code = "";
+
+        while (isNot) {
+            code = randomString.generate({
+                length: 6,
+                charset: "alphanumeric"
+            });
+
+            const user: User | null = await User.findOne({
+                where: { code }
+            });
+
+            if (!user) isNot = false;
+        }
+
+        return code;
+    };
+
+    private createProfilePath = (userId: number, file: File): string | null => {
+        let path: string | null = "";
+        const reqFileName = file.originalFilename!;
+        const isDefault = isDefaultFile(reqFileName);
+
+        /**
+         * Frontend에선 static으로 default.jpg,png,svg 셋 중 하나 갖고있다가
+         * 사용자가 profile을 내리면 그걸로 넣고 요청
+         */
+        if (isDefault) path = null;
+        else path = `${this.FOLDER_NAME}/${userId}/profile/${dayjs().valueOf()}.${reqFileName}`;
+
+        return path;
+    };
+
     getUser = async (userId: number): Promise<IUserResponse> => {
         const user1: User | null = await User.findOne({
             attributes: { exclude: ["password"] },
             where: { userId }
         });
+
         let user2: User | null = null;
 
         if (!user1) throw new UnauthorizedError("User not found with given ID");
@@ -29,10 +77,113 @@ class UserService {
 
         const result: IUserResponse = {
             ...user1.dataValues,
-            couple: user2
+            couple: user2 ? user2 : undefined
         };
 
         return result;
+    };
+
+    /**
+     * 유저 정보를 생성합니다.
+     * @param data A {@link User}
+     */
+    createUser = async (transaction: Transaction, data: ICreate): Promise<void> => {
+        const user: User | null = await User.findOne({
+            where: {
+                [Op.or]: [{ email: data.email }, { phone: data.phone }]
+            }
+        });
+
+        if (user) throw new ConflictError("Duplicated User");
+
+        const hash: string = await createDigest(data.password);
+        data.password = hash;
+        data.code = await this.createCode();
+
+        const createdUser: User = await User.create(
+            {
+                snsId: data.snsId,
+                code: data.code,
+                name: data.name,
+                email: data.email,
+                birthday: new Date(data.birthday),
+                password: hash,
+                phone: data.phone,
+                eventNofi: data.eventNofi
+            },
+            { transaction }
+        );
+
+        await UserRole.create(
+            {
+                userId: createdUser.userId,
+                roleId: 4
+            },
+            { transaction }
+        );
+    };
+
+    updateUser = async (transaction: Transaction, data: IUpdate, file?: File): Promise<User> => {
+        let isUpload = false;
+        let path: string | null = "";
+
+        const user: User | null = await User.findOne({
+            where: { userId: data.userId }
+        });
+
+        if (!user) throw new NotFoundError("Not Found User");
+        else if (user.deleted) throw new ForbiddenError("User is deleted");
+
+        let prevProfile: string | null = user.profile;
+
+        try {
+            if (file) {
+                data.profile = this.createProfilePath(data.userId, file);
+
+                // profile 있으면 업로드
+                if (data.profile) {
+                    await uploadFile(data.profile, file.filepath);
+                    isUpload = true;
+
+                    if (prevProfile) await deleteFile(prevProfile); // 전에 있던 profile 삭제
+                } else if (prevProfile && !data.profile) {
+                    // default 이미지로 변경시
+                    await deleteFile(prevProfile);
+                }
+            }
+
+            await user.update(data);
+
+            return user;
+        } catch (error) {
+            // Firebase에는 업로드 되었지만 DB 오류가 발생했다면 Firebase Profile 삭제
+            if (data.profile && isUpload) {
+                await deleteFile(path!);
+                logger.error(`After updating the firebase, a db error occurred and the firebase profile is deleted => ${path}`);
+            }
+
+            throw error;
+        }
+    };
+
+    /**
+     * 사용자 정보 삭제이며, Couple이 있는 경우 Frontend에서 연인 끊기 후 삭제를 요청.
+     * @param userId User Id
+     */
+    deleteUser = async (transaction: Transaction, userId: number): Promise<void> => {
+        const user: User | null = await User.findOne({ where: { userId: userId } });
+
+        if (!user) throw new NotFoundError("Not Found User");
+
+        await user.update(
+            {
+                deleted: true,
+                deletedTime: new Date(dayjs().valueOf())
+            },
+            { transaction }
+        );
+
+        logger.debug(`Success Deleted userId => ${userId}`);
     };
 }
